Switch slicing procedures to take int64 indices
This has two major benefits:

- We can consistently raise a range-error? when being passed a negative
  or large index. Right now we will raise a type-error? if we can
  statically determine it's negative or truncate otherwise. We should be
  checking integer ranges before truncating but in this case we're
  checking the indices anyway so we can simply do it at their full 64bit
  size

- This will make it easier to support 64bit lengthed bytevectors on
  64bit platforms.

This also adds missing functional tests for some ranged procedures. In
particular, (bytevector-copy) wasn't validating its range at all and is
fixed to call assertSliceValid as part of this work.

+++ runtime/stdlib/scheme/base/input.cpp
@@ -200,7 +200,7 @@ AnyCell *llbase_read_bytevector(World &world, std::uint32_t requestedBytes, Port
 	return BytevectorCell::withByteArray(world, byteArray, readBytes);
 }
 
-AnyCell *llbase_mutating_read_bytevector(World &world, BytevectorCell *bytevector, PortCell *portCell, std::uint32_t start, std::uint32_t end)
+AnyCell *llbase_mutating_read_bytevector(World &world, BytevectorCell *bytevector, PortCell *portCell, std::int64_t start, std::int64_t end)
 {
 	if (bytevector->isGlobalConstant())
 	{
