+++ runtime/stdlib/scheme/base/output.cpp
@@ -46,7 +46,7 @@ void llbase_write_char(World &world, UnicodeChar character, PortCell *portCell)
 	portStream->write(reinterpret_cast<char*>(utf8Bytes.data), utf8Bytes.size);
 }
 
-void llbase_write_string(World &world, StringCell *stringCell, PortCell *portCell, std::uint32_t start, std::uint32_t end)
+void llbase_write_string(World &world, StringCell *stringCell, PortCell *portCell, std::int64_t start, std::int64_t end)
 {
 	assertSliceValid(world, "(write-string)", stringCell, stringCell->charLength(), start, end);
 
@@ -58,7 +58,7 @@ void llbase_write_string(World &world, StringCell *stringCell, PortCell *portCel
 	portStream->write(reinterpret_cast<const char *>(range.startPointer), range.endPointer - range.startPointer);
 }
 
-void llbase_write_bytevector(World &world, BytevectorCell *bytevectorCell, PortCell *portCell, std::uint32_t start, std::uint32_t end)
+void llbase_write_bytevector(World &world, BytevectorCell *bytevectorCell, PortCell *portCell, std::int64_t start, std::int64_t end)
 {
 	assertSliceValid(world, "(write-bytevector)", bytevectorCell, bytevectorCell->length(), start, end);
